fix missing query var
