+++ server/package.json
@@ -15,7 +15,10 @@
     ]
   },
   "jest": {
-    "testEnvironment": "node"
+    "testEnvironment": "node",
+    "testMatch": [
+      "**/__tests__/**/*.test.js"
+    ]
   },
   "scripts": {
     "start": "yarn dev",
